Add checks before doing any operations in database

+++ src/Flights.Data/Database/Repositories/FlightQueryNotificationRepository.cs
@@ -36,6 +36,9 @@ public class FlightQueryNotificationRepository : IFlightQueryNotificationReposit
 
     public async Task DeleteNotificationsAsync(FlightQueryNotification[] notifications)
     {
+        if (notifications.Length == 0) 
+            return;
+
         var notificationIds = notifications.Select(x => x.Id).ToArray();
 
         await collection.DeleteManyAsync(n => notificationIds.Contains(n.Id));
